move search bar to molecules
